+++ tests/test_main.py
@@ -2,14 +2,15 @@ from unittest import TestCase
 
 import chess
 
-from chess_server import main, chessgame
+from chess_server import main
 from chess_server.main import RESPONSES
+from chess_server.utils import User
 
 
 class TestGetResultComment(TestCase):
     def test_get_result_comment_game_is_not_over_as_white(self):
 
-        user = chessgame.User(board=chess.Board(), color=chess.WHITE)
+        user = User(board=chess.Board(), color=chess.WHITE)
 
         expected = None
         result = main.get_result_comment(user)
@@ -20,7 +21,7 @@ class TestGetResultComment(TestCase):
 
         # Reference game: Morphy-Maurian, 1869
 
-        user = chessgame.User(
+        user = User(
             board=chess.Board(
                 "r1bq2k1/ppp2Bp1/2np1n1p/b3p3/3PP3/B1P5/P4PPP/RN1Q1RK1 b - - "
                 "0 11"
@@ -35,7 +36,7 @@ class TestGetResultComment(TestCase):
 
     def test_get_result_comment_user_wins_as_white(self):
 
-        user = chessgame.User(
+        user = User(
             board=chess.Board("4N3/ppp3Qk/2p5/7p/3P3P/8/P4P2/5K2 b - - 0 36"),
             color=chess.WHITE,
         )
@@ -47,7 +48,7 @@ class TestGetResultComment(TestCase):
 
     def test_get_result_comment_user_wins_as_black(self):
 
-        user = chessgame.User(
+        user = User(
             board=chess.Board(
                 "rnb1kbnr/pppp1ppp/8/4p3/6Pq/5P2/PPPPP2P/RNBQKBNR w KQkq - 1 3"
             ),
@@ -64,7 +65,7 @@ class TestGetResultComment(TestCase):
         expected = RESPONSES["result_draw"].format(reason="stalemate")
 
         # As white
-        user = chessgame.User(
+        user = User(
             board=chess.Board("7K/7P/7k/8/6q1/8/8/8 w - - 0 1"),
             color=chess.WHITE,
         )
@@ -73,7 +74,7 @@ class TestGetResultComment(TestCase):
         self.assertEqual(result, expected)
 
         # As black
-        user = chessgame.User(
+        user = User(
             board=chess.Board("7K/7P/7k/8/6q1/8/8/8 w - - 0 1"),
             color=chess.BLACK,
         )
@@ -88,7 +89,7 @@ class TestGetResultComment(TestCase):
         )
 
         # As white
-        user = chessgame.User(
+        user = User(
             board=chess.Board("4k3/8/8/8/8/5B2/8/4K3 w - - 0 1"),
             color=chess.WHITE,
         )
@@ -97,7 +98,7 @@ class TestGetResultComment(TestCase):
         self.assertEqual(result, expected)
 
         # As black
-        user = chessgame.User(
+        user = User(
             board=chess.Board("4k3/8/8/8/8/5B2/8/4K3 w - - 0 1"),
             color=chess.BLACK,
         )
@@ -110,7 +111,7 @@ class TestGetResultComment(TestCase):
         expected = RESPONSES["result_draw"].format(reason="fifty move rule")
 
         # As white
-        user = chessgame.User(
+        user = User(
             board=chess.Board("4k3/8/6r1/8/8/8/2R5/4K3 w - - 120 1"),
             color=chess.WHITE,
         )
@@ -119,7 +120,7 @@ class TestGetResultComment(TestCase):
         self.assertEqual(result, expected)
 
         # As black
-        user = chessgame.User(
+        user = User(
             board=chess.Board("4k3/8/6r1/8/8/8/2R5/4K3 w - - 120 1"),
             color=chess.BLACK,
         )
@@ -135,7 +136,7 @@ class TestGetResultComment(TestCase):
 
         # As white
         board = chess.Board()
-        user = chessgame.User(board=board, color=chess.WHITE)
+        user = User(board=board, color=chess.WHITE)
 
         self.assertEqual(main.get_result_comment(user=user), None)
         board.push_san("Nf3")
@@ -155,7 +156,7 @@ class TestGetResultComment(TestCase):
 
         # As black
         board = chess.Board()
-        user = chessgame.User(board=board, color=chess.BLACK)
+        user = User(board=board, color=chess.BLACK)
 
         board.push_san("Nf3")
         self.assertEqual(main.get_result_comment(user=user), None)
